+++ src/main/java/com/progra3/proyecto/repository/RepartidorRepository.java
@@ -6,8 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.progra3.proyecto.entity.Repartidor;
 
-public interface RepartidorRepository extends JpaRepository<Repartidor, Integer>{
-	
-	List<Repartidor> findByNombre(String nombre);
-	
+public interface RepartidorRepository extends JpaRepository<Repartidor, Long>{
+	List<Repartidor> findByNombre(String nombre);	
 }
