+++ src/main/java/com/progra3/proyecto/service/jpa/RepartidorServiceImpl.java
@@ -21,13 +21,12 @@ public class RepartidorServiceImpl implements IRepartidorService {
 		return repo.findAll();
 	}
 	
-
     @Override
     public List<Repartidor> buscarRepartidor(String nombre) {
         return repo.findByNombre(nombre);
     }
 	@Override
-	public Repartidor getById(Integer id) {
+	public Repartidor getById(Long id) {
 		Optional<Repartidor> optional = repo.findById(id);
 		return optional.orElse(null);
 	}
@@ -38,7 +37,7 @@ public class RepartidorServiceImpl implements IRepartidorService {
 	}
 
 	@Override
-	public void delete(Integer id) {
+	public void delete(Long id) {
 		Optional<Repartidor> optional = repo.findById(id);
 		if (optional.isPresent()) {
 			repo.deleteById(id);
@@ -46,7 +45,7 @@ public class RepartidorServiceImpl implements IRepartidorService {
 	}
 
 	@Override
-	public boolean exists(Integer id) {
+	public boolean exists(Long id) {
 		return (id == null) ? false : repo.existsById(id);
 	}
